RF: Move statistics BioScan dataset into bioscan_statistic.py
